+++ hi-ml/testhiml/testhiml/test_run_ml.py
@@ -1,18 +1,17 @@
+import shutil
 from pathlib import Path
 
 import pytest
-from typing import Tuple
-from unittest.mock import patch, MagicMock, Mock
+from typing import Generator, Tuple
+from unittest.mock import patch, MagicMock
 
 from health_ml.configs.hello_container import HelloContainer
-from pytorch_lightning import Callback
-
 from health_ml.experiment_config import ExperimentConfig
 from health_ml.lightning_container import LightningContainer
 from health_ml.run_ml import MLRunner
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def ml_runner_no_setup() -> MLRunner:
     experiment_config = ExperimentConfig(model="HelloContainer")
     container = LightningContainer(num_epochs=1)
@@ -20,41 +19,39 @@ def ml_runner_no_setup() -> MLRunner:
     return runner
 
 
-@pytest.fixture
-def ml_runner() -> MLRunner:
+@pytest.fixture(scope="module")
+def ml_runner() -> Generator:
     experiment_config = ExperimentConfig(model="HelloContainer")
     container = LightningContainer(num_epochs=1)
     runner = MLRunner(experiment_config=experiment_config, container=container)
     runner.setup()
-    return runner
+    yield runner
+    output_dir = runner.container.file_system_config.outputs_folder
+    if output_dir.exists():
+        shutil.rmtree(output_dir)
 
 
-@pytest.fixture
-def ml_runner_with_container() -> MLRunner:
+@pytest.fixture(scope="module")
+def ml_runner_with_container() -> Generator:
     experiment_config = ExperimentConfig(model="HelloContainer")
     container = HelloContainer()
     runner = MLRunner(experiment_config=experiment_config, container=container)
     runner.setup()
-    return runner
+    yield runner
+    output_dir = runner.container.file_system_config.outputs_folder
+    if output_dir.exists():
+        shutil.rmtree(output_dir)
 
 
-# @pytest.fixture
-# def ml_runner() -> MLRunner:
-#     experiment_config = ExperimentConfig(model="HelloContainer")
-#     container = LightningContainer(num_epochs=1)
-#     runner = MLRunner(experiment_config=experiment_config, container=container)
-#     runner.setup()
-#     return runner
+def _mock_model_train(chekpoint_path: Path, container: LightningContainer) -> Tuple[str, str]:
+    return "trainer", "storing_logger"
 
 
 def test_ml_runner_setup(ml_runner_no_setup: MLRunner) -> None:
-    """
-    Check that all the necessary methods get called during setup
-    """
+    """Check that all the necessary methods get called during setup"""
     assert not ml_runner_no_setup._has_setup_run
     with patch.object(ml_runner_no_setup, "container", spec=LightningContainer) as mock_container:
         with patch("health_ml.run_ml.seed_everything") as mock_seed:
-            # mock_container.get_effectie_random_seed = Mock()
             ml_runner_no_setup.setup()
             mock_container.get_effective_random_seed.assert_called_once()
             mock_container.setup.assert_called_once()
@@ -64,6 +61,7 @@ def test_ml_runner_setup(ml_runner_no_setup: MLRunner) -> None:
 
 
 def test_set_run_tags_from_parent(ml_runner: MLRunner) -> None:
+    """Test that set_run_tags_from_parents causes set_tags to get called"""
     with pytest.raises(AssertionError) as ae:
         ml_runner.set_run_tags_from_parent()
         assert "should only be called in a Hyperdrive run" in str(ae)
@@ -75,14 +73,10 @@ def test_set_run_tags_from_parent(ml_runner: MLRunner) -> None:
             mock_run_context.set_tags.assert_called()
 
 
-def _mock_model_train(chekpoint_path: Path, container: LightningContainer) -> Tuple[str, str]:
-    return "trainer", "storing_logger"
-
-
 def test_run(ml_runner: MLRunner) -> None:
     """Test that model runner gets called """
     ml_runner.setup()
-    assert not ml_runner.has_continued_training
+    assert not ml_runner.checkpoint_handler.has_continued_training
     with patch.object(ml_runner, "run_inference_for_lightning_models"):
         with patch.object(ml_runner, "checkpoint_handler"):
             with patch("health_ml.run_ml.model_train", new=_mock_model_train):
@@ -91,45 +85,14 @@ def test_run(ml_runner: MLRunner) -> None:
                 # expect _mock_model_train to be called and the result of ml_runner.storing_logger
                 # updated accordingly
                 assert ml_runner.storing_logger == "storing_logger"
-                assert ml_runner.has_continued_training
+                assert ml_runner.checkpoint_handler.has_continued_training
 
 
-@patch("health_ml.run_ml.create_lightning_trainer")
-def test_run_inference_for_lightning_models(mock_create_trainer: MagicMock, ml_runner: MLRunner,
-                                            tmp_path: Path) -> None:
+def test_run_inference_for_lightning_models(ml_runner_with_container: MLRunner) -> None:
     """
-    Check that all expected methods are called during inference3
+    Test that run_inference_for_lightning_models gets called as expected. If no checkpoint paths are
+    provided, should raise an error. Otherwise, expect the trainer object's test method to be called
     """
-    mock_trainer = MagicMock()
-    mock_test_result = [{"result": 1.0}]
-    mock_trainer.test.return_value = mock_test_result
-    mock_create_trainer.return_value = mock_trainer, ""
-
-    with patch.object(ml_runner, "container") as mock_container:
-        mock_container.num_gpus_per_node.return_value = 0
-        mock_container.get_trainer_arguments.return_value = {"callbacks": Callback()}
-        mock_container.load_model_checkpoint.return_value = Mock()
-        mock_container.get_data_module.return_value = Mock()
-        mock_container.pl_progress_bar_refresh_rate = None
-        mock_container.detect_anomaly = False
-        mock_container.pl_limit_train_batches = 1.0
-        mock_container.pl_limit_val_batches = 1.0
-        mock_container.outputs_folder = tmp_path
-
-        checkpoint_paths = [Path("dummy")]
-        result = ml_runner.run_inference_for_lightning_models(checkpoint_paths)
-        assert result == mock_test_result
-
-        mock_create_trainer.assert_called_once()
-        mock_container.load_model_checkpoint.assert_called_once()
-        mock_container.get_data_module.assert_called_once()
-        mock_trainer.test.assert_called_once()
-
-
-def test_on_test_epoch_end(ml_runner_with_container: MLRunner):
-    # def _mock_load_model_checkpoint(checkpoint_path):
-    #     ml_runner._model = ml_runner.model
-
     with patch.object(ml_runner_with_container, "checkpoint_handler") as mock_checkpoint_handler:
         with patch("health_ml.run_ml.model_train", new=_mock_model_train):
             with pytest.raises(ValueError) as e:
@@ -139,8 +102,6 @@ def test_on_test_epoch_end(ml_runner_with_container: MLRunner):
         with patch.object(ml_runner_with_container.container, "load_model_checkpoint"):
             with patch("health_ml.model_trainer.create_lightning_trainer") as mock_create_train_trainer:
                 with patch("health_ml.run_ml.create_lightning_trainer") as mock_create_inference_trainer:
-                    # set the model which currently is None
-                    # ml_runner.container._model = HelloRegression()
                     mock_trainer = MagicMock()
                     mock_create_train_trainer.return_value = mock_trainer, ""
                     mock_create_inference_trainer.return_value = mock_trainer, ""